requirements.txt
absl-py==2.1.0
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
awscli==1.40.31
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.0
deepface==0.0.93
dlib==19.24.6
docutils==0.19
easydict==1.13
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
fastapi==0.115.11
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.0
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
prettytable==3.16.0
protobuf==5.29.3
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
rsa==4.7.2
s3transfer==0.13.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
